Build multiple bundles automatically.
